requirements.txt
accelerate==1.7.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
authlib==1.3.1
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colpali-engine==0.3.10
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.58.0
fqdn==1.5.1
fsspec==2025.5.1
gputil==1.4.0
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.2
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook-shim==0.2.4
numpy==1.26.4
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
prometheus-client==0.21.1
prompt-toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
safetensors==0.5.3
scipy==1.15.3
send2trash==1.8.3
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.13.2
typing-inspection==0.4.0
uri-template==1.3.0
urllib3==2.4.0
validators==0.34.0
wcwidth==0.2.13
weaviate-client @ git+https://github.com/weaviate/weaviate-python-client.git@014f0a3b1e49025f6cba9aa2e11c09b4cc989a66
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
